requirements.txt
numpy
opencv-python

# For GPIO access (rotary encoder)
RPi.GPIO; platform_machine == "armv7l"
fake-rpi; platform_machine != "armv7l"

